Add verifyService and sanitize search
Refactor middleware functions to be organized better and added middleware to sanitize user search input
