+++ backend/services/authService.js
@@ -94,51 +94,3 @@ exports.AuthorizeUser = async (req, res, next) => {
         return res.status(403).json({ message: "Invalid or expired token. Please log in again." });
     }
 };
-
-function validateId(id) {
-    if(!id.match(/^[0-9a-fA-F]{24}$/)) {
-        let err = new Error('Invalid id: ' + id);
-        err.status = 400;
-        return err;
-    } else {
-        return null;
-    }
-}
-
-/**
- * Middleware function to verify passed lastId is correctly formatted
- */
-exports.VerifyLastId = (req, res, next) => {
-
-    try {
-        // Grab the ?lastId=, if doesn't exist, null.
-        let id = req.query.lastId || null;
-        if(!id) return next(); // If no id, no error needed.
-
-        const err = validateId(id);
-        if(err) {
-            return next(err);
-        } else {
-            return next();
-        }
-    }
-    catch (err) { next(err); }
-};
-
-/**
- * Middleware function to verify parameter id is correctly formatted
- */
-exports.VerifyParamsId = (req, res, next) => {
-    try {
-        // Grab the /:id
-        let id = req.params.id;
-
-        const err = validateId(id);
-        if(err) {
-            return next(err);
-        } else {
-            return next();
-        }
-    }
-    catch (err) { next(err); }
-}
